Add files via upload
The data consists of customers divided into calibration and validation sub-samples contained in the data file called "cell2cell.dta". The calibration data contain customers, approximately half of whom churn. The rest of them are in the validation data, with a small portion of them churn. The roughly 50% churn rate in the calibration data is not realistic, but the "oversampling" of churners makes it easier to identify the profile of churners as distinct from the profile of non-churners. The validation data contain a small percentage of churners, which is the current monthly churn rate at Cell2Cell.

General Instructions
• This case requires you to develop a model for predicting customer churn at "Cell2Cell," a fictitious wireless telecom company, and use insights from the model to develop an incentive plan for enticing would-be churners to remain with Cell2Cell.
• The data for the case are a scaled-down version of the full database generously donated by an anonymous wireless telephone company. There are over 65 potential predictors. Assume that logistic regression is the best choice to develop your predictive model.
• This case requires both statistical analysis and creativity/judgment. I recommend that you not spend too much time fine-tuning your predictive model; make sure you spend enough time interpreting and using the results.